# Optimize worst db query